Add CustomListView & uploadReview func

+++ app/src/main/java/com/song2/boostcourse/data/ReviewData.java
@@ -36,7 +36,7 @@ public class ReviewData {
         return rate;
     }
 
-    public void setRate(int rate) {
+    public void setRate(float rate) {
         this.rate = rate;
     }
